Squashed 'Externals/Core/' changes from d95ecf3..ba2b9f4
ba2b9f4 Corrected property type in DTO
72db898 Correction to parameter type in contract
9e7408a Merge branch 'master' of http://git.mothershipdev.com:7990/scm/ms2/mothershipv2_core into TaskNotification
8b54a78 Merge pull request #28 in MS2/mothershipv2_core from Offers to master
27db9a9 type of transaction change for iridiumrequest
86b1a59 updated enums
2fc5f13 updated enums
7dd726f removed duplicate event status key property
27263c0 Merge branch 'master' into TaskNotification
3502bd6 added config value for imap
dff3d77 Refactor IEchoSignCallbackService to IEchoSignPushService
f17c234 Sync Core with Services and Web
45fb456 Merge branch 'master' into TaskNotification
42be2cf renamed 2 properties
ca54eec Merge branch 'master' into TaskNotification
0fce786 Merge branch 'master' into TaskNotification
0dad837 updated dto
fc96397 commit display name
6d482be added display name to notification icon enum
5505046 added notification enum
3cf7229 changed the message callback service
b689cb6 added totalNotificationCount to pushes
26d12d6 added class to notification received
ebd9dad new send msg service
603993b added new dto to hold statuses and notifications
4d68494 Merge branch 'master' into TaskNotification
a481c51 added userguid sendmessagerequestdto
05394fc update to eventnoteslnk
ab7eac7 update to event role dto
db7bcf9 updated event dto
9cfe7ac updated dto structure to mirror that of what is in the entity framework
94959e5 removed Quartz from task notification
843f23d add execute job to dialer task notification
3acd680 move quartz into domain
4047cf2 Merge branch 'master' into TaskNotification
43e642c removed message listener from send message service
7cff25d added update date to status owner dto
ea26c6c change send message service receive sms to use inboundMessage
9bff002 add InboundMessage to core
15ea0ac Merge branch 'master' into TaskNotification
3bd3188 changes to the dto structure
fb1b62e improving the service caching
a8fe576 updating the message service status subscription
84f5ea5 change username property in received notification
6a65fb9 Added usernames property to notificationReceivedDto
1d957c8 added subscribe methods to the sendmessage interface
bcaa79f update to dto name in queue email message
1819f3b Merge branch 'master' into TaskNotification
e5a5b0d removed quartzjob from send message service
facf156 add user guid to send message service
4e0ee7b contract for get event notifications
e5b8045 rename svc
77e034a rename dto property
ac56744 update to send msg callback
b1d0255 Message notification push callback
d68e6f3 updates to push service status
d849423 message service status dto updated
7167311 update to send message service
cf3b5e3 Merge branch 'master' into TaskNotification
e6a9e53 undo last comit
80eaba3 service constract added to INotificationService
2a6f501 Added INotification interface
97b90f6 callback refactoring for task notification
df1aa94 Update to dtos to use enums
d9ba5f3 Removed notification guid from task dto
bd6805d Merge branch 'master' into TaskNotification
9a74827 change date time format
b690046 Return data from service
23a65e3 delete changed to close
c06f3b5 Added canceltask to task service
894a36a renames
2c3f377 Refactoring to not delete or update
b30133f added delete task to ITaskNotificationService
1151e8c Merge branch 'master' into TaskNotification
2e805e5 get task list change for the date - sql
f60a35e dto changes update, save and get
66a5ff6 Dto Name change
f40fa7b unsaved dto
8a6a24c Organising Dtos for notifications
7d3cc69 Merge branch 'master' into TaskNotification
33c6f9c added notification dto
5d61381 adding notifications to the db with an event

git-subtree-dir: Externals/Core
git-subtree-split: ba2b9f4dacddb7ec3061b19df26dd10864fef103

+++ PS.Mothership.Core/PS.Mothership.Core.Common/Contracts/IAddressManagementService.cs
@@ -20,7 +20,7 @@ namespace PS.Mothership.Core.Common.Contracts
         ICollection<CountryDto> Countries();
 
         [OperationContract]
-        ICollection<CountyDto> Counties(long countryKey);
+        ICollection<CountyDto> Counties(int countryKey);
 
         [OperationContract]
         IEnumerable<PostcodeDto> Postcodes(FilterDto filter);
